+++ src/main/java/com/smartsheet/api/internal/AbstractResources.java
@@ -528,7 +528,9 @@ public abstract class AbstractResources {
      * @throws SmartsheetRestException : if there is any other REST API related error occurred during the operation
      * @throws SmartsheetException : if there is any other error occurred during the operation
      */
-    protected <T> TokenPaginatedResult<T> listResourcesWithTokenPagination(String path, JsonDeserializer<List<T>> deserializer) throws SmartsheetException {
+    protected <T> TokenPaginatedResult<T> listResourcesWithTokenPagination(String path,
+                                                                            JsonDeserializer<List<T>> deserializer)
+            throws SmartsheetException {
         Util.throwIfNull(path, deserializer);
         Util.throwIfEmpty(path);