+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [x.x.x] - unreleased
+
 ## [3.5.0] - 2025-08-07
 ### Added
   - Added support for token-based pagination in WorkspaceResources.listWorkspaces() method